# docs/swagger.yaml
basePath: /
definitions:
  GetPackageByIdRow:
    properties:
      description:
        type: string
      drug_id:
        type: integer
      eans:
        items:
          type: string
        type: array
      id:
        type: integer
      presentation_registration:
        type: string
      updated_at:
        type: string
    type: object
...
  ListPackagesRow:
    properties:
      description:
        type: string
      drug_id:
        type: integer
      eans:
        items:
          type: string
        type: array
      id:
        type: integer
      presentation_registration:
        type: string
      updated_at:
        type: string
    type: object
...
  createPackageRequest:
    properties:
      description:
        example: 400 mg, caixa com 20 cápsulas
        type: string
      eans:
        example:
        - "7891058001155"
        - "7891058017392"
        items:
          type: string
        type: array
      presentation_registration:
        example: "1234567890014"
        type: string
      registration_number:
        example: "1234567890123"
        type: string
    type: object
...
  idResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  listResponse:
    properties:
      data: {}
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  updatePackageRequest:
    properties:
      description:
        example: 400 mg, caixa com 20 cápsulas
        type: string
      drug_id:
        example: 1
        type: integer
      eans:
        example:
        - "7891058001155"
        - "7891058017392"
        items:
          type: string
        type: array
      presentation_registration:
        example: "1234567890014"
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    API that reads the barcode on a medicine box and returns the simplified leaflet.
    Public route: GET /drugs/ean/{ean}. All others require login (POST /auth/login sets a session cookie).
  title: PharmaCode API
  version: "1.0"
paths:
...
  /packages:
    get:
      parameters:
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: How many items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/listResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ListPackagesRow'
                  type: array
              type: object
        "400":
          description: invalid limit / invalid offset
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: List packages (paginated)
      tags:
      - packages
    post:
      consumes:
      - application/json
      description: A package can have several EANs (e.g. an old and a new code on
        the shelf at the same time).
      parameters:
      - description: Package data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createPackageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/idResponse'
        "400":
          description: invalid json or missing required field
          schema:
            type: string
        "404":
          description: drug not found
          schema:
            type: string
        "409":
          description: ean or presentation_registration already registered
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a package (presentation) of a drug
      tags:
      - packages
  /packages/{id}:
    delete:
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: package not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete a package
      tags:
      - packages
    get:
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetPackageByIdRow'
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: package not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get a package by ID
      tags:
      - packages
    put:
      consumes:
      - application/json
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: integer
      - description: Package data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updatePackageRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id or json
          schema:
            type: string
        "404":
          description: package not found / drug not found
          schema:
            type: string
        "409":
          description: ean or presentation_registration already registered
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update a package
      tags:
      - packages
...
swagger: "2.0"